+++ src/main/java/ma/nttdata/externals/module/interview/dto/InterviewEvaluationsRequestDTO.java
@@ -3,7 +3,6 @@ package ma.nttdata.externals.module.interview.dto;
 import java.util.List;
 
 public record InterviewEvaluationsRequestDTO(
-        String promptCode,
         List<QuestionsAndAnswersForEvaluationDTO> questionsAndAnswersForEvaluation
 ) {
 }
